+++ frontend/src/i18n/home.ts
@@ -256,6 +256,14 @@ export default {
     en: 'Creator',
     fr: 'Créateur·rice',
   },
+  planner_table_years_tco2eq: {
+    en: 'Per year (t CO₂-eq)',
+    fr: 'Par année (t CO₂-eq)',
+  },
+  planner_table_grant_tco2eq: {
+    en: 'Grant proposal (t CO₂-eq)',
+    fr: 'Demande de financement (t CO₂-eq)',
+  },
   planner_table_action: {
     en: 'Action',
     fr: 'Action',